T2I/requirements.txt
# 文本生成图片幻觉检测与消除项目依赖

# 基础依赖
torch>=2.6.0
torchvision>=0.19.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
safetensors>=0.4.0

# 图像处理
Pillow>=9.0.0
opencv-python>=4.7.0
scikit-image>=0.20.0

# 评估指标相关
ftfy>=6.1.1
regex>=2022.7.9
sentence-transformers>=2.2.0
lpips>=0.1.4

# 数据处理
datasets>=2.12.0
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# 可视化（服务器兼容）
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# 系统监控（可选）
psutil>=5.9.0

# 工具库
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# HTTP请求
requests>=2.28.0
urllib3>=1.26.0

# 机器学习
scikit-learn>=1.2.0
xgboost>=1.7.0

# 多模态相关
open-clip-torch>=2.20.0
timm>=0.9.0

# YAML配置文件支持
PyYAML>=6.0

# 强化学习（阶段三需要）
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# 检索增强（阶段三需要）
faiss-cpu>=1.7.4
chromadb>=0.4.0
